# trie: reduce hasher allocations