--- a/trie/hasher.go
+++ b/trie/hasher.go
@@ -56,7 +56,7 @@ func (b *sliceBuffer) Reset() {
 var hasherPool = sync.Pool{
 	New: func() interface{} {
 		return &hasher{
-			tmp: make(sliceBuffer, 0, 256),
+			tmp: make(sliceBuffer, 0, 550), // cap is as large as a full fullNode.
 			sha: sha3.NewKeccak256().(keccakState),
 		}
 	},